server/requirements.txt
fastapi[all]
uvicorn
google-cloud-firestore
